# swagger.yaml
openapi: 3.0.3
info:
  title: Mock API - Tunnel Reservation
  version: 2.0.0
  description: >
    API REST mock Node.js + Express pour simuler Snaplogic pendant le developpement.
    Architecture simplifiee : chaque GET /residences/:id retourne la fiche complete
    (typologies, scenarios pricing, offres, admin overlay) en un seul appel.
servers:
  - url: http://localhost:8081
    description: Local mock server
...
paths:
...
  /cities/{cityAlias}/residences:
    get:
      tags: [Residences]
      summary: GetResidencesByCity
      description: >
        Residences d'une ville (filtre sur `cityAlias` ou `city`).
        Ex. `/cities/Paris/residences` retourne les 4 ECLA.
      parameters:
        - in: path
          name: cityAlias
          required: true
          schema:
            type: string
          description: Alias de ville (ex. Paris, Lyon, Toulouse).
      responses:
        "200":
          description: Residences de la ville
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CityResidencesResponse"
  /cities/{cityAlias}/residences/{id}:
    get:
      tags: [Residences]
      summary: GetOneResidence
      description: >
        Fiche residence complete en un seul appel.
        Inclut `typologies` (pricing + amenities par typologie), `configurationOptions`
        (grilles d’options unitaires communes a la residence), `offerSummaries`, `offers` et `offersContext`.
        Le front charge et cache cette reponse pour alimenter tout le tunnel.
      parameters:
        - in: path
          name: cityAlias
          required: true
          schema:
            type: string
          description: Alias de ville (alimente les etats in-app).
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: UUID de la residence (ex. 19f2179b-7d14-f011-998a-6045bd1919a1).
      responses:
        "200":
          description: Fiche residence complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResidenceDetail"
        "404":
          description: Residence introuvable
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  cityAlias:
                    type: string
                  residenceId:
                    type: string
  /admin-tr:
    get:
      tags: [AdminTR]
      summary: GetInformationAdminTR
      description: >
        Configuration du tunnel de reservation geree par l'espace admin.
        Sans query param : retourne uniquement `modals` + `steps` (config globale, boot).
        Avec `?city=` : retourne en plus `residenceOverrides` — flags marketing et photos
        filtrees sur la ville demandee (match sur `city` ou `cityAlias`).
      parameters:
        - in: query
          name: city
          required: false
          schema:
            type: string
          description: >
            Alias de ville (ex. Paris, Lyon). Si fourni, la reponse inclut
            `residenceOverrides` avec les flags et photos des residences de cette ville.
      responses:
        "200":
          description: Configuration admin tunnel (+ overrides si city fourni)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminTRResponse"
  /offers:
    get:
      tags: [Offers]
      summary: GetOffers
      description: >
        Referentiel offres global (vue Fabric / outils). Pour le tunnel React,
        privilegier `offerSummaries` et `offers` sur GET /residences et GET /residences/:id.
      parameters:
        - in: query
          name: city
          required: false
          schema:
            type: string
        - in: query
          name: bookingDate
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Liste des offres
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OffersResponse"
  /reservations:
    post:
      tags: [Reservations]
      summary: Creation opportunite ECCO
      description: |
        Simule la creation d'une opportunite dans le CRM ECCO.
        Le query param `scenario` selectionne le fichier mock a retourner.
        Defaut : `locataire_seul_majeur`.
      parameters:
        - in: query
          name: scenario
          required: false
          schema:
            type: string
            enum:
              - locataire_seul_majeur
              - locataire_seul_mineur_garant_physique
          description: Scenario de creation (defaut locataire_seul_majeur).
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Opportunite creee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpportunityCreationResponse"
components:
  schemas:
    Coordinates:
      type: object
      properties:
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
      required: [lat, lng]
    TypologyPreview:
      type: object
      description: >
        Apercu typologie sur le listing. La ligne colocation peut utiliser `Sous_type`
        + bornes `MinimumBaseRent*` au lieu de `typologyCode` + `lowestUnitPrice`.
      properties:
        typologyCode:
          type: string
          example: STUDIO
          nullable: true
        Sous_type:
          type: string
          nullable: true
          description: Libelle sous-type pour la ligne colocation (ex. Colocation).
        sous-type-marketing_fr:
          type: string
          description: Libelle marketing FR affiche sur le listing.
        sous_type_marketing_en:
          type: string
          description: Libelle marketing EN affiche sur le listing.
        lowestUnitPrice:
          type: number
          nullable: true
          description: >
            Borne basse du loyer pour cette typologie (hors ligne colocation `Sous_type`).
        MinimumBaseRent:
          type: number
          nullable: true
        MinimumBaseRentWithOptions:
          type: number
          nullable: true
        MinimumBaseRentWithOptionsWithOffers:
          type: number
          nullable: true
        reducedBaseRent:
          type: number
          description: Loyer reduit eventuel (promo), si fourni par la vue.
      required:
        - sous-type-marketing_fr
        - sous_type_marketing_en
    Floors:
      type: object
      properties:
        min:
          type: integer
        max:
          type: integer
      required: [min, max]
    ResidenceTag:
      type: object
      nullable: true
      properties:
        code:
          type: string
          enum: [POPULAR, SPECIAL_OFFER, NEW, LAST_UNITS]
        label:
          type: string
      required: [code, label]
    ResidencePhoto:
      type: object
      properties:
        url:
          type: string
          format: uri
        alt:
          type: string
        order:
          type: integer
        context:
          type: string
          enum: [HERO, COMMON, TYPOLOGY]
        typologyCode:
          type: string
          nullable: true
          description: Present uniquement si context = TYPOLOGY.
      required: [url, alt, order, context]
    ResidenceListItem:
      type: object
      properties:
        residenceId:
          type: string
        residenceCode:
          type: string
        residenceCodeAnalytic:
          type: string
          description: Code analytique / comptable (ERP) associe a la residence.
        name:
          type: string
        commercialName:
          type: string
          description: Nom commercial de la residence (peut differer du nom technique).
        brand:
          type: string
          enum: [ECLA, UXCO STUDENT]
        city:
          type: string
        cityAlias:
          type: string
          description: Alias de ville pour le front (ECLA => Paris).
        address:
          type: string
        postal_code:
          type: string
          description: Code postal (5 chiffres).
        coordinates:
          $ref: "#/components/schemas/Coordinates"
        Residence_availability:
          type: string
          description: >
            Disponibilite synthetique (ex. `AVAILABLE || FULL`, `PARTIAL || FULL`).
        MinimumBaseRent:
          type: number
        MinimumBaseRentWithOptions:
          type: number
        MinimumBaseRentWithOptionsWithOffers:
          type: number
        availabilityStatus:
          type: string
          deprecated: true
          description: Remplace par `Residence_availability` dans le mock listing.
        baseRentFrom:
          type: number
          deprecated: true
          description: Remplace par `MinimumBaseRent*` dans le mock listing.
        tag:
          $ref: "#/components/schemas/ResidenceTag"
          description: >
            Absent du fichier masterdata ; le serveur mock l’ajoute si un `flag`
            existe dans `GetAdminTR.residenceOverrides` pour cette residence.
        typologyTags:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ResidenceTag"
          description: >
            Idem : injecte par merge admin (heritage du flag sur les typologies ciblees).
        photos:
          type: array
          items:
            $ref: "#/components/schemas/ResidencePhoto"
        typologies:
          type: array
          items:
            $ref: "#/components/schemas/TypologyPreview"
        offerSummaries:
          type: array
          items:
            $ref: "#/components/schemas/OfferSummary"
      required:
        - residenceId
        - residenceCode
        - residenceCodeAnalytic
        - name
        - commercialName
        - brand
        - city
        - cityAlias
        - address
        - postal_code
        - coordinates
        - photos
        - Residence_availability
        - MinimumBaseRent
        - MinimumBaseRentWithOptions
        - MinimumBaseRentWithOptionsWithOffers
        - typologies
        - offerSummaries
    OffersContext:
      type: object
      description: Metadonnees communes au referentiel offres (alignement cache).
      properties:
        bookingDate:
          type: string
          format: date
        city:
          type: string
      required: [bookingDate, city]
    OfferSummary:
      type: object
      properties:
        offerCode:
          type: string
        label:
          type: string
        badge:
          type: string
          nullable: true
        type:
          type: string
        promoCode:
          type: string
          nullable: true
        bookingDeadline:
          type: string
          format: date
          nullable: true
        discountTarget:
          type: string
          nullable: true
        discountAmount:
          type: number
          nullable: true
        discountReductionType:
          type: string
          nullable: true
        billingTiming:
          type: string
          nullable: true
      required: [offerCode, label, type]
...
    CityResidencesResponse:
      type: object
      description: Residences d'une ville (GET /cities/{cityAlias}/residences).
      properties:
        cityAlias:
          type: string
        offersContext:
          $ref: "#/components/schemas/OffersContext"
        residences:
          type: array
          items:
            $ref: "#/components/schemas/ResidenceListItem"
      required: [cityAlias, offersContext, residences]
    CommonAmenity:
      type: object
      properties:
        code:
          type: string
        label:
          type: string
      required: [code, label]
    AbonnementResidence:
      type: object
      description: >
        Service recurrent optionnel (ménage, parking, location TV, pack) affiche sur la fiche residence.
      properties:
        code:
          type: string
          description: Identifiant fonctionnel (ex. CLEANING_SERVICE, PARKING_SPACE).
        title:
          type: string
        description:
          type: string
        pricePerMonth:
          type: number
        currency:
          type: string
          example: EUR
        billingPeriod:
          type: string
          description: Periode de facturation (ex. month).
        icon:
          type: string
          description: Cle d’icone UI (ex. cleaning, parking, tv, gift).
      required:
        - code
        - title
        - description
        - pricePerMonth
        - currency
        - billingPeriod
        - icon
    TypologyDetail:
      type: object
      description: >
        Typologie sur la fiche residence : tarification et amenities inclus logement.
        Les options configurables (etage, surface, etc.) sont dans `configurationOptions`.
      properties:
        typologyCode:
          type: string
        label:
          type: string
        surface:
          type: number
          description: Surface indicative (m2), si fournie par la vue.
        reducedBaseRent:
          type: number
          description: Loyer reduit eventuel (promo).
        pricing:
          $ref: "#/components/schemas/Pricing"
        amenities:
          type: array
          items:
            $ref: "#/components/schemas/Amenity"
      required:
        - typologyCode
        - label
        - pricing
        - amenities
    Pricing:
      type: object
      description: >
        Bornes tarifaires par typologie (loyer, options, offres, frais et caution).
        `MinimumBaseRentWithOptionsWithOffers` utilise `reducedBaseRent` de la typologie
        lorsqu’il est strictement inferieur au loyer de base, sinon loyer de base - 20 €.
      properties:
        MinimumBaseRent:
          type: number
        MinimumBaseRentWithOptions:
          type: number
        MinimumBaseRentWithOptionsWithOffers:
          type: number
        MinimumApplicationFee:
          type: number
        MinimumSecurityDeposit:
          type: number
        MinimumApplicationFeeWithOffers:
          type: number
        MinimumSecurityDepositWithOffers:
          type: number
      required:
        - MinimumBaseRent
        - MinimumBaseRentWithOptions
        - MinimumBaseRentWithOptionsWithOffers
        - MinimumApplicationFee
        - MinimumSecurityDeposit
        - MinimumApplicationFeeWithOffers
        - MinimumSecurityDepositWithOffers
    Amenity:
      type: object
      properties:
        code:
          type: string
        label:
          type: string
      required: [code, label]
    OptionVariant:
      type: object
      properties:
        variantCode:
          type: string
        label:
          type: string
        description:
          type: string
          nullable: true
        amount:
          type: number
        billingTiming:
          type: string
        isDefault:
          type: boolean
        productCode:
          type: string
          nullable: true
        accountingDestinationCode:
          type: string
          nullable: true
      required: [variantCode, label, amount, billingTiming, isDefault]
    OptionGroup:
      type: object
      description: >
        Groupe d’options configurables ; pas d’icone en masterdata (mapping UI cote client ou referentiel separe).
      properties:
        groupCode:
          type: string
        label:
          type: string
        variants:
          type: array
          items:
            $ref: "#/components/schemas/OptionVariant"
      required: [groupCode, label, variants]
    ResidenceDetail:
      type: object
      description: >
        Fiche complete residence : infos generales, typologies (pricing + amenities),
        options unitaires communes (`configurationOptions`), offres completes et contexte offres.
        Un seul appel pour le tunnel.
      properties:
        residenceId:
          type: string
        residenceCode:
          type: string
        name:
          type: string
        commercialName:
          type: string
        brand:
          type: string
          enum: [ECLA, UXCO STUDENT]
        city:
          type: string
        cityAlias:
          type: string
        address:
          type: string
        telephone:
          type: string
          description: Numero de contact residence (affichage / click-to-call).
        email:
          type: string
          format: email
          description: Adresse e-mail de contact residence.
        coordinates:
          $ref: "#/components/schemas/Coordinates"
        availabilityStatus:
          type: string
        description:
          type: string
        tag:
          $ref: "#/components/schemas/ResidenceTag"
          description: >
            Absent du fichier masterdata ; le serveur mock l’ajoute si un `flag`
            existe dans `GetAdminTR.residenceOverrides` pour cette residence.
        typologyTags:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ResidenceTag"
          description: >
            Idem : injecte par merge admin (heritage du flag sur les typologies ciblees).
        photos:
          type: array
          items:
            $ref: "#/components/schemas/ResidencePhoto"
        commonAmenities:
          type: array
          items:
            $ref: "#/components/schemas/CommonAmenity"
        abonnements:
          type: array
          items:
            $ref: "#/components/schemas/AbonnementResidence"
        typologies:
          type: array
          items:
            $ref: "#/components/schemas/TypologyDetail"
        floors:
          $ref: "#/components/schemas/Floors"
        configurationOptions:
          type: array
          description: >
            Options configurables communes a la residence (etage, surface, literie, clim, balcon, etc.).
            Les montants sont identiques quel que soit le type de logement selectionne.
            Le groupe `PREMIUM` n’est present que pour les residences qui le proposent (ex. Noisy-le-Grand).
          items:
            $ref: "#/components/schemas/OptionGroup"
        offersContext:
          $ref: "#/components/schemas/OffersContext"
        offerSummaries:
          type: array
          items:
            $ref: "#/components/schemas/OfferSummary"
        offers:
          type: array
          items:
            $ref: "#/components/schemas/Offer"
      required:
        - residenceId
        - residenceCode
        - name
        - commercialName
        - brand
        - city
        - cityAlias
        - address
        - telephone
        - email
        - coordinates
        - availabilityStatus
        - description
        - photos
        - commonAmenities
        - abonnements
        - typologies
        - floors
        - configurationOptions
        - offersContext
        - offerSummaries
        - offers
    OfferScope:
      type: object
      properties:
        residences:
          type: array
          items:
            type: string
        typologies:
          type: array
          items:
            type: string
      required: [residences, typologies]
    OfferConditions:
      type: object
      properties:
        bookingDeadline:
          type: string
          format: date
          nullable: true
        startDateFrom:
          type: string
          format: date
          nullable: true
        startDateTo:
          type: string
          format: date
          nullable: true
        minDurationMonths:
          type: integer
          nullable: true
      required: [bookingDeadline, startDateFrom, startDateTo, minDurationMonths]
    OfferDiscount:
      type: object
      properties:
        target:
          type: string
        reductionType:
          type: string
        amount:
          type: number
        billingTiming:
          type: string
        recurrenceMonths:
          type: integer
          nullable: true
      required: [target, reductionType, amount, billingTiming]
    OfferAccounting:
      type: object
      properties:
        productCode:
          type: string
        accountingDestinationCode:
          type: string
      required: [productCode, accountingDestinationCode]
    Offer:
      type: object
      properties:
        offerCode:
          type: string
        label:
          type: string
        badge:
          type: string
          nullable: true
        promoCode:
          type: string
          nullable: true
        scope:
          $ref: "#/components/schemas/OfferScope"
        conditions:
          $ref: "#/components/schemas/OfferConditions"
        discount:
          $ref: "#/components/schemas/OfferDiscount"
        accounting:
          $ref: "#/components/schemas/OfferAccounting"
        type:
          type: string
      required:
        - offerCode
        - label
        - promoCode
        - scope
        - conditions
        - discount
        - accounting
        - type
    OffersResponse:
      type: object
      properties:
        bookingDate:
          type: string
          format: date
        city:
          type: string
        offers:
          type: array
          items:
            $ref: "#/components/schemas/Offer"
      required: [bookingDate, city, offers]
...
    ModalLocalized:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: string
        sections:
          type: array
          items:
            $ref: "#/components/schemas/ModalSection"
      required: [title, subtitle, sections]
    Modal:
      type: object
      properties:
        step:
          type: integer
        trigger:
          type: string
          description: Evenement declencheur (ex. EXPRESS_BOOKING_CLICK, PAYMENT_STEP_ENTER).
        fr:
          $ref: "#/components/schemas/ModalLocalized"
        en:
          $ref: "#/components/schemas/ModalLocalized"
      required: [step, trigger, fr, en]
    TunnelStepLocalized:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
      required: [title, description]
    TunnelStep:
      type: object
      properties:
        stepCode:
          type: string
        order:
          type: integer
        fr:
          $ref: "#/components/schemas/TunnelStepLocalized"
        en:
          $ref: "#/components/schemas/TunnelStepLocalized"
      required: [stepCode, order, fr, en]
    AdminCommonAmenity:
      type: object
      description: Equipement commun residence (FR + EN) pour la surcouche admin.
      properties:
        code:
          type: string
        label:
          type: string
          description: Libelle affiche (FR).
        label_en:
          type: string
          description: Libelle anglais.
      required: [code, label, label_en]
    ResidenceOverrideFlag:
      type: object
      description: >
        Flag marketing attribue via l'admin. Heritage parent-enfant :
        si `typologies` est vide ou absent, toutes les typologies de la residence heritent du flag.
        Si `typologies` est renseigne, seules ces typologies le portent.
      properties:
        code:
          type: string
          enum: [POPULAR, SPECIAL_OFFER, NEW, LAST_UNITS]
        label:
          type: string
        label_en:
          type: string
          description: Libelle du flag en anglais (tunnel bilingue).
        typologies:
          type: array
          items:
            type: string
          description: >
            Codes typo cibles (ex. STUDIO, T2). Vide = toutes les typologies heritent.
      required: [code, label]
    ResidenceOverride:
      type: object
      description: >
        Surcouche admin pour une residence : texte marketing, equipements communs,
        flag marketing et photos. Les photos remplacent integralement celles de la masterdata.
      properties:
        residenceId:
          type: string
        city:
          type: string
        cityAlias:
          type: string
        description:
          type: string
          description: Texte descriptif marketing (tunnel / fiche enrichie).
        commonAmenities:
          type: array
          items:
            $ref: "#/components/schemas/AdminCommonAmenity"
          description: Equipements communs avec libelles FR et EN.
        flag:
          $ref: "#/components/schemas/ResidenceOverrideFlag"
        photos:
          type: array
          items:
            $ref: "#/components/schemas/ResidencePhoto"
      required: [residenceId, city, cityAlias]
    AdminTRResponse:
      type: object
      description: >
        Configuration du tunnel de reservation.
        Modals bilingues par step+trigger, steps informatifs.
        `residenceOverrides` n'est present que si le query param `city` est fourni.
      properties:
        modals:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/Modal"
        steps:
          type: array
          items:
            $ref: "#/components/schemas/TunnelStep"
        residenceOverrides:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ResidenceOverride"
          description: >
            Cle = residenceId (UUID). Present uniquement si query param `city` fourni.
            Contient les flags et photos admin des residences de la ville demandee.
      required: [modals, steps]
...
    OpportunityScenarioPayload:
      type: object
      description: |
        Payload de creation opportunite ECCO.
        `contacts[]` est dynamique : seuls les contacts renseignes sont inclus.
        Chaque contact porte un `uxc_role` (picklist ECCO) qui determine son role.
        Les garants ne sont pas des contacts — ils sont captures via
        `necessiteDeGarant` et `typeDeGarant` sur l opportunity.
      properties:
        contacts:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/Contact"
        opportunity:
          $ref: "#/components/schemas/OpportunityDetail"
      required: [contacts, opportunity]
    OpportunityCreationResponse:
      type: object
      properties:
        status:
          type: string
          example: ACCEPTED
        opportunityId:
          type: string
          example: OPP-M1A2B3C4
        scenario:
          type: string
          example: locataire_seul_majeur
        availableScenarios:
          type: array
          items:
            type: string
        submittedPayload:
          $ref: "#/components/schemas/OpportunityScenarioPayload"
        confirmation:
          $ref: "#/components/schemas/ReservationConfirmation"
      required: [status, opportunityId, scenario, availableScenarios, submittedPayload, confirmation]
    ReservationConfirmation:
      type: object
      properties:
        status:
          type: string
          example: ACCEPTED
        quoteRequestId:
          type: string
          example: BQ-2026-000123
        eccoOpportunityId:
          type: string
          example: OPP-2026-XXXXX
        message:
          type: string
          example: Reservation creee avec succes dans ECCO.